changed User.newest_tweet_id and Tweet.id to numeric datatypes, to support IDs up to 20 digits in length

+++ twitoff/models.py
@@ -8,14 +8,14 @@ DB = SQLAlchemy()
 class User( DB.Model):
 	id = DB.Column( DB.BigInteger, primary_key= True)
 	name = DB.Column( DB.String( 20), nullable= False)
-	newest_tweet_id = DB.Column( DB.BigInteger)
+	newest_tweet_id = DB.Column( DB.Numeric( 20, 0))
 
 	def __repr__( self):
 		return '<User {}>'.format( self.name)
 
 
 class Tweet( DB.Model):
-	id = DB.Column( DB.BigInteger, primary_key= True)
+	id = DB.Column( DB.Numeric( 20, 0), primary_key= True)
 	text = DB.Column( DB.Unicode( 500), nullable= False)
 	embedding = DB.Column( DB.PickleType, nullable= False)
 	user_id = DB.Column( DB.BigInteger, DB.ForeignKey( 'user.id'), nullable= False)
